requirements_v2.txt
numpy==2.2.6
pandas==2.3.3
pandas-ta==0.4.71b0
yfinance
python-dotenv
pyyaml
pyarrow
requests
plotly
duckdb
matplotlib
pydantic
schedule
tabulate
